third_party/blink/web_tests/external/wpt/webrtc-stats/rtp-stats-creation-expected.txt

This is a testharness.js-based test.
[FAIL] No RTCRtpStreamStats exist when only remote description is set
  assert_equals: no rtp stats with only remote description expected 0 but got 2
[FAIL] No RTCRtpStreamStats exist when only local description is set
  assert_equals: no rtp stats with only local description expected 0 but got 2
[FAIL] No RTCOutboundRtpStreamStats exist until packets have been sent
  assert_true: no outbound rtp stats before packets sent expected true got false
[FAIL] No RTCInboundRtpStreamStats exist until packets have been received
  assert_true: no inbound rtp stats before packets received expected true got false
Harness: the test ran to completion.
